Fix config objects beeing non existing on startup in REDIS tests (#2025)
* promisify start/stopController

* rm hack

* allow typescript for tests

* fix linter

* own tsconfig for tests

* port from require to import for setup4controller

* prettify test

* add back pack lock
