+++ lerna.json
@@ -1,17 +1,22 @@
 {
-    "useWorkspaces": true,
-    "version": "5.0.0-alpha.0-20221202-deec19ee",
-    "//": "Set up some command options",
-    "command": {
-        "//": "Running a script should show the output while it is generated",
-        "run": {
-            "stream": true
-        },
-        "version": {
-            "//": "This is required for the release-script",
-            "allowBranch": ["master", "*.x"],
-            "ignoreChanges": ["*.md"],
-            "exact": true
-        }
+  "useWorkspaces": true,
+  "version": "5.0.0-alpha.0-20221202-deec19ee",
+  "//": "Set up some command options",
+  "command": {
+    "//": "Running a script should show the output while it is generated",
+    "run": {
+      "stream": true
+    },
+    "version": {
+      "//": "This is required for the release-script",
+      "allowBranch": [
+        "master",
+        "*.x"
+      ],
+      "ignoreChanges": [
+        "*.md"
+      ],
+      "exact": true
     }
+  }
 }
