+++ package.json
@@ -25,6 +25,7 @@
     "@alcalzone/release-script-plugin-license": "~3.5.0",
     "@alcalzone/release-script-plugin-manual-review": "~3.5.8",
     "@tsconfig/node12": "^1.0.11",
+    "@types/chai": "^4.3.3",
     "@types/debug": "^4.1.7",
     "@types/event-stream": "^4.0.0",
     "@types/fs-extra": "^9.0.13",
@@ -62,6 +63,7 @@
     "semver": "^7.3.7",
     "sinon": "^15.0.0",
     "sinon-chai": "^3.7.0",
+    "ts-node": "^10.9.1",
     "typedoc": "^0.23.11",
     "typedoc-plugin-markdown": "^3.13.5",
     "typedoc-plugin-missing-exports": "^1.0.0",
@@ -77,9 +79,9 @@
     "release": "release-script",
     "lint": "eslint .",
     "prettier": "eslint . --fix",
-    "test": "mocha packages/controller/test --exit",
-    "test-jsonl": "mocha packages/controller/test/jsonl/ --exit",
-    "test-redis-socket": "mocha packages/controller/test/redis-socket/ --exit",
-    "test-redis-sentinel": "mocha packages/controller/test/redis-sentinel/ --exit"
+    "test": "mocha -r ts-node/register packages/controller/test/*.ts --exit",
+    "test-jsonl": "mocha -r ts-node/register packages/controller/test/jsonl/*.ts --exit",
+    "test-redis-socket": "mocha -r ts-node/register packages/controller/test/redis-socket/*.ts --exit",
+    "test-redis-sentinel": "mocha -r ts-node/register packages/controller/test/redis-sentinel/*.ts --exit"
   }
 }
