+++ packages/controller/test/.eslintrc
@@ -1,3 +1,6 @@
 env:
   node: true
   mocha: true
+parserOptions:
+  sourceType: module
+  project: './tsconfig.json'
