+++ packages/controller/test/lib/testObjectsACL.js
@@ -121,17 +121,15 @@ function register(it, expect, context) {
             );
     }).timeout(2000);
 
-    it(textName + 'invalid user name must be checked #1', () => {
+    it(textName + 'invalid user name must be checked #1', async () => {
         const objects = context.objects;
-        return objects
-            .getObject(secretId, { user: 'admin' })
-            .then(_obj => {
-                expect(1).to.be.equal('Never happens');
-            })
-            .catch(err => {
-                console.error(err.message);
-                expect(err.message).to.be.equal('permissionError');
-            });
+        try {
+            await objects.getObject(secretId, { user: 'admin' });
+            expect(1).to.be.equal('Never happens');
+        } catch (e) {
+            console.error(e.message);
+            expect(e.message).to.be.equal('permissionError');
+        }
     }).timeout(2000);
 
     it(textName + 'invalid user name must be checked #2', () => {
@@ -174,15 +172,6 @@ function register(it, expect, context) {
 
     it(textName + 'default acl from system.config should be used', async () => {
         const objects = context.objects;
-
-        // hack for redis, because testing setup missing setup first logic
-        objects.defaultNewAcl = {
-            object: 1636,
-            state: 1636,
-            file: 1636,
-            owner: 'system.user.governor',
-            ownerGroup: 'system.group.senatorGroup'
-        };
 
         await objects.setObjectAsync('test.defAck', { type: 'state' });
         const obj = await objects.getObjectAsync('test.defAck');
