+++ packages/controller/test/testAdapterStatesRedis.ts
@@ -1,10 +1,5 @@
-/* jshint -W097 */
-/* jshint strict:false */
-/* jslint node:true */
-/* jshint expr:true */
-'use strict';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const testAdapter = require('./lib/testAdapter');
-// const dataDir = __dirname + '/../tmp/data-redis';
 
 const statesConfig = {
     options: {
