+++ packages/controller/test/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  // own json config, for test files root is different, just using ts-node no build
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "."
+  },
+  "include": [
+    "**/*.ts"
+  ],
+  "exclude": [
+  ]
+}
